Update TSL237__Arduino_Uno_Compatible_.ino
I realized that I accidentally had >= for the modulo operation, which in the context of this code guaranteed it to always be true. Because the code has worked fine under such circumstances, I decided to remove the code that was only relevant if the conditions of the final if statement could be false.