Added CSV-formatted code for Arduino
For easier data analysis, I decided to create another version of the Arduino code with the serial print statements formatted so that they can be saved as a CSV file. I created this as a separate file so that we have the option to jump between the .txt and .csv versions whenever needed.