DS18B20 replaces TMP36 temperature sensor
The code was altered to accommodate the more accurate DS18B20 temperature sensor. The sensor runs via the DallasTemperature library and OneWire. To my knowledge, only digital pins 2 and 4 work with the sensor's data pin, but let me know if anything else will allow the DS18B20 to transmit data through OneWire. This sensor not only improves accuracy but decreases the amount of data spit out, meaning alterations to the ARDread program on the pi will have to be made in the near future.